+++ app/repositories/migrations/versions/V0008__wallpapers_bing_portrait_image_url.sql
@@ -1,0 +1,2 @@
+ALTER TABLE wallpapers
+    ADD COLUMN portrait_image_url TEXT;